// src/app/core/api/model/topologyAppDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TopologyAppReplicaDto } from './topologyAppReplicaDto';


export interface TopologyAppDto { 
    id: string;
    name: string;
    /**
     * Pre-computed short label, max 14 chars
     */
    slug: string;
    displayName: string;
    category: TopologyAppDto.CategoryEnum;
    kind: TopologyAppDto.KindEnum;
    namespace: string;
    status: TopologyAppDto.StatusEnum;
    statusReason: string | null;
    ramRequestMB: number;
    ramLimitMB: number;
    /**
     * CPU request in millicores
     */
    cpuRequestM: number;
    /**
     * CPU limit in millicores
     */
    cpuLimitM: number;
    /**
     * Server hosting the leader/primary planet for rendering
     */
    primaryServerId: string;
    replicas: Array<TopologyAppReplicaDto>;
    /**
     * Sum of replicas[].count, denormalized for the client
     */
    replicaCount: number;
    scalingMode: TopologyAppDto.ScalingModeEnum;
    scalingNote?: string | null;
    labels: { [key: string]: string; };
}
export namespace TopologyAppDto {
    export const CategoryEnum = {
        Database: 'database',
        Cache: 'cache',
        Storage: 'storage',
        Automation: 'automation',
        Media: 'media',
        Monitoring: 'monitoring',
        Web: 'web',
        Business: 'business',
        Infra: 'infra'
    } as const;
    export type CategoryEnum = typeof CategoryEnum[keyof typeof CategoryEnum];
    export const KindEnum = {
        User: 'user',
        System: 'system'
    } as const;
    export type KindEnum = typeof KindEnum[keyof typeof KindEnum];
    export const StatusEnum = {
        Running: 'running',
        Warning: 'warning',
        Error: 'error',
        Stopped: 'stopped'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const ScalingModeEnum = {
        Manual: 'manual',
        Hpa: 'hpa',
        Daemonset: 'daemonset',
        Statefulset: 'statefulset'
    } as const;
    export type ScalingModeEnum = typeof ScalingModeEnum[keyof typeof ScalingModeEnum];
}